+++ docs/Whats-New.md
@@ -540,6 +540,7 @@ New:
 - Added amount limit of `LimboKill` (by NetsuNegi)
 - [Customizations for techno type target scan/guard range](Fixed-or-Improved-Logics.md#target-scan-guard-range-customizations) (by Starkku)
 - [Spawns particle when spawns tiberium by terrain](Fixed-or-Improved-Logics.md#customizable-ore-spawners) (by NetsuNegi)
+- Allow jumpjet climbing ignore building height (by TaranDahl)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)